Build Age Calculator App
In this we will make an Age Calculator app with the help of HTML, CSS and JavaScript. This age calculator app has one data input box, where it display the calendar to pick date of birth. after selecting date of birth click on calculate button then it will display the exact age till today.
This age calculator app will display the exact age in years, months and days.
#JavaScriptProject